Relocate classes.
