+++ deb/openmediavault/usr/lib/python3/dist-packages/openmediavault/device/block.py
@@ -20,118 +20,12 @@
 # along with OpenMediaVault. If not, see <http://www.gnu.org/licenses/>.
 import os
 import re
-import stat
 import openmediavault.subprocess
 
-
-def is_block_device(path):
-    """
-    Check if path is a device file, e.g. /dev/sda1.
-    :param path: The path to check.
-    :return: Return True if path describes a device file,
-        otherwise False.
-    :rtype: bool
-    """
-    if not isinstance(path, str):
-        return False
-    if not os.path.exists(path):
-        return False
-    return stat.S_ISBLK(os.stat(path).st_mode)
-
-
-def is_device_file(path):
-    """
-    Check if path describes a device file, e.g. /dev/sda1.
-    :param path: The path to check.
-    :return: Return True if path describes a device file,
-        otherwise False.
-    :rtype: bool
-    """
-    if not isinstance(path, str):
-        return False
-    return True if re.match(r'^\/dev\/.+$', path) else False
-
-
-def is_device_file_by(path):
-    """
-    Check if path describes a device file, e.g.
-
-    * /dev/disk/by-uuid/ad3ee177-777c-4ad3-8353-9562f85c0895
-    * /dev/disk/by-id/usb-Kingston_DataTraveler_G2_001CC0EC21ADF011C6A20E35-0:0-part1
-    * /dev/disk/by-label/data
-
-    :param path: The path to check.
-    :return: Return True if path describes a device file,
-        otherwise False.
-    :rtype: bool
-    """
-    if not isinstance(path, str):
-        return False
-    return True if re.match(r'^\/dev\/disk\/by-\S+\/.+$', path) else False
-
-
-def is_device_file_by_uuid(path):
-    """
-    Check if path describes a device file, e.g.
-
-    * /dev/disk/by-uuid/ad3ee177-777c-4ad3-8353-9562f85c0895
-    * /dev/disk/by-uuid/2ED43920D438EC29 (NTFS)
-
-    :param path: The path to check.
-    :return: Return True if path describes a device file,
-        otherwise False.
-    :rtype: bool
-    """
-    if not isinstance(path, str):
-        return False
-    return True if re.match(r'^\/dev\/disk\/by-uuid\/.+$', path) else False
-
-
-def is_device_file_by_id(path):
-    """
-    Check if path describes a device file, e.g.
-
-    * /dev/disk/by-id/usb-Kingston_DataTraveler_G2_001CC0EC21ADF011C6A20E35-0:0-part1
-    * /dev/disk/by-id/wwn-0x4002c554a4d79cb9-part2
-
-    :param path: The path to check.
-    :return: Return True if path describes a device file,
-        otherwise False.
-    :rtype: bool
-    """
-    if not isinstance(path, str):
-        return False
-    return True if re.match(r'^\/dev\/disk\/by-id\/.+$', path) else False
-
-
-def is_device_file_by_label(path):
-    """
-    Check if path describes a device file, e.g. /dev/disk/by-label/data
-    :param path: The path to check.
-    :return: Return True if path describes a device file,
-        otherwise False.
-    :rtype: bool
-    """
-    if not isinstance(path, str):
-        return False
-    return True if re.match(r'^\/dev\/disk\/by-label\/.+$', path) else False
-
-
-def is_device_file_by_path(path):
-    """
-    Check if path describes a device file, e.g.
-
-    * /dev/disk/by-path/pci-0000:00:17.0-ata-3
-    * /dev/disk/by-path/pci-0000:00:17.0-ata-3-part2
-
-    :param path: The path to check.
-    :return: Return True if path describes a device file,
-        otherwise False.
-    :rtype: bool
-    """
-    if not isinstance(path, str):
-        return False
-    return True if re.match(r'^\/dev\/disk\/by-path\/.+$', path) else False
+from .utils import (
+    is_block_device, is_device_file, is_device_file_by_id,
+    is_device_file_by_path
+)
 
 
 class BlockDevice:
